Added royal advisor sprite

+++ Gone Is The King/Assets/Scripts/Character Scripts/IAttackable.cs
@@ -2,7 +2,7 @@ using UnityEngine;
 
 public interface IAttackable
 {
-    public int Health {get;set;}
-    public int Defense {get;set;}
-    public void TakeDamage(int dmg);
+    public double Health {get;set;}
+    public double Defense {get;set;}
+    public void TakeDamage(double dmg);
 }
